android: Update gradle/dbflow and other dependencies

+++ android/app/src/main/java/be/justcode/bandtracker/activity/BandDetailsActivity.java
@@ -182,7 +182,7 @@ public class BandDetailsActivity extends AppCompatActivity {
 
         @Override
         public int getItemCount() {
-            return mCursor.getCount();
+            return (int) mCursor.getCount();
         }
 
         public void rowClicked(int position) {
