+++ android/app/src/main/java/be/justcode/bandtracker/activity/MainBandsFragment.java
@@ -115,7 +115,7 @@ public class MainBandsFragment extends Fragment {
 
         @Override
         public int getItemCount() {
-            return mCursor.getCount();
+            return (int) mCursor.getCount();
         }
 
         public void refresh() {
